# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MenuStack - Restaurant Ordering & POS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  /** Process uptime in seconds */
  uptime: number;
  /** Database connection status (ok or error) */
  database: string;
  /** ISO 8601 timestamp of the health check */
  timestamp: string;
}

export interface MenuCategory {
  id: number;
  name: string;
  description?: string | null;
  sortOrder: number;
  createdAt: string;
}

export interface CreateMenuCategoryBody {
  name: string;
  description?: string | null;
  sortOrder: number;
}

export interface MenuItem {
  id: number;
  categoryId: number;
  name: string;
  description?: string | null;
  price: number;
  isAvailable: boolean;
  imageUrl?: string | null;
  createdAt: string;
}

export interface CreateMenuItemBody {
  categoryId: number;
  name: string;
  description?: string | null;
  price: number;
  isAvailable: boolean;
  imageUrl?: string | null;
}

export interface MenuCategoryWithItems {
  id: number;
  name: string;
  description?: string | null;
  sortOrder: number;
  items: MenuItem[];
}

export interface OrderItem {
  id: number;
  orderId: number;
  menuItemId: number;
  menuItemName: string;
  quantity: number;
  unitPrice: number;
}

export type OrderType = (typeof OrderType)[keyof typeof OrderType];

export const OrderType = {
  online: "online",
  in_person: "in_person",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  new: "new",
  preparing: "preparing",
  ready: "ready",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type OrderPaymentStatus =
  (typeof OrderPaymentStatus)[keyof typeof OrderPaymentStatus];

export const OrderPaymentStatus = {
  unpaid: "unpaid",
  paid: "paid",
} as const;

export type OrderPaymentMethod =
  | (typeof OrderPaymentMethod)[keyof typeof OrderPaymentMethod]
  | null;

export const OrderPaymentMethod = {
  cash: "cash",
  card: "card",
  online: "online",
} as const;

export interface Order {
  id: number;
  type: OrderType;
  customerName?: string | null;
  tableNumber?: number | null;
  status: OrderStatus;
  paymentStatus: OrderPaymentStatus;
  paymentMethod?: OrderPaymentMethod;
  total: number;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type OrderWithItems = Order & {
  items: OrderItem[];
};

export type CreateOrderBodyType =
  (typeof CreateOrderBodyType)[keyof typeof CreateOrderBodyType];

export const CreateOrderBodyType = {
  online: "online",
  in_person: "in_person",
} as const;

export interface CreateOrderItemBody {
  menuItemId: number;
  quantity: number;
}

export interface CreateOrderBody {
  type: CreateOrderBodyType;
  customerName?: string | null;
  tableNumber?: number | null;
  notes?: string | null;
  items: CreateOrderItemBody[];
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  new: "new",
  preparing: "preparing",
  ready: "ready",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export type UpdateOrderPaymentBodyPaymentMethod =
  (typeof UpdateOrderPaymentBodyPaymentMethod)[keyof typeof UpdateOrderPaymentBodyPaymentMethod];

export const UpdateOrderPaymentBodyPaymentMethod = {
  cash: "cash",
  card: "card",
  online: "online",
} as const;

export interface UpdateOrderPaymentBody {
  paymentMethod: UpdateOrderPaymentBodyPaymentMethod;
}

export interface DashboardStats {
  ordersToday: number;
  revenueToday: number;
  avgOrderValue: number;
  activeOrders: number;
  topSellingCategory?: string | null;
  paidOrders: number;
  unpaidOrders: number;
}

export interface TopSellingItem {
  menuItemId: number;
  name: string;
  totalQuantity: number;
  totalRevenue: number;
}

export interface RevenueByHour {
  hour: number;
  revenue: number;
  orderCount: number;
}

export interface OrderStatusCount {
  status: string;
  count: number;
}

export type ListMenuItemsParams = {
  categoryId?: number;
  available?: boolean;
};

export type ListOrdersParams = {
  status?: ListOrdersStatus;
  type?: ListOrdersType;
  paymentStatus?: ListOrdersPaymentStatus;
  /**
   * ISO date string to filter by day (YYYY-MM-DD)
   */
  date?: string;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  new: "new",
  preparing: "preparing",
  ready: "ready",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type ListOrdersType =
  (typeof ListOrdersType)[keyof typeof ListOrdersType];

export const ListOrdersType = {
  online: "online",
  in_person: "in_person",
} as const;

export type ListOrdersPaymentStatus =
  (typeof ListOrdersPaymentStatus)[keyof typeof ListOrdersPaymentStatus];

export const ListOrdersPaymentStatus = {
  unpaid: "unpaid",
  paid: "paid",
} as const;

export type GetTopSellingItemsParams = {
  limit?: number;
};